--- a/HPPH.Benchmark/AverageBenchmarks.cs
+++ b/HPPH.Benchmark/AverageBenchmarks.cs
@@ -13,6 +13,8 @@ public class AverageBenchmarks
 
     private readonly List<ColorRGB[]> _colors3bpp;
     private readonly List<ColorRGBA[]> _colors4bpp;
+    private readonly List<IImage<ColorRGB>> _images3bpp;
+    private readonly List<IImage<ColorRGBA>> _images4bpp;
 
     #endregion
 
@@ -22,6 +24,9 @@ public class AverageBenchmarks
     {
         _colors3bpp = BenchmarkHelper.GetSampleData<ColorRGB>();
         _colors4bpp = BenchmarkHelper.GetSampleData<ColorRGBA>();
+
+        _images3bpp = BenchmarkHelper.GetSampleDataImages<ColorRGB>();
+        _images4bpp = BenchmarkHelper.GetSampleDataImages<ColorRGBA>();
     }
 
     #endregion
@@ -44,6 +49,26 @@ public class AverageBenchmarks
         ColorRGBA[] averages = new ColorRGBA[_colors4bpp.Count];
         for (int i = 0; i < _colors4bpp.Count; i++)
             averages[i] = new ReadOnlySpan<ColorRGBA>(_colors4bpp[i]).Average();
+
+        return averages;
+    }
+
+    [Benchmark]
+    public ColorRGB[] PixelHelper_3BPP_Image()
+    {
+        ColorRGB[] averages = new ColorRGB[_images3bpp.Count];
+        for (int i = 0; i < _images3bpp.Count; i++)
+            averages[i] = _images3bpp[i].Average();
+
+        return averages;
+    }
+
+    [Benchmark]
+    public ColorRGBA[] PixelHelper_4BPP_Image()
+    {
+        ColorRGBA[] averages = new ColorRGBA[_images4bpp.Count];
+        for (int i = 0; i < _images4bpp.Count; i++)
+            averages[i] = _images4bpp[i].Average();
 
         return averages;
     }