--- a/HPPH.Benchmark/MinMaxBenchmarks.cs
+++ b/HPPH.Benchmark/MinMaxBenchmarks.cs
@@ -13,6 +13,8 @@ public class MinMaxBenchmarks
 
     private readonly List<ColorRGB[]> _colors3bpp;
     private readonly List<ColorRGBA[]> _colors4bpp;
+    private readonly List<IImage<ColorRGB>> _images3bpp;
+    private readonly List<IImage<ColorRGBA>> _images4bpp;
 
     #endregion
 
@@ -22,6 +24,9 @@ public class MinMaxBenchmarks
     {
         _colors3bpp = BenchmarkHelper.GetSampleData<ColorRGB>();
         _colors4bpp = BenchmarkHelper.GetSampleData<ColorRGBA>();
+
+        _images3bpp = BenchmarkHelper.GetSampleDataImages<ColorRGB>();
+        _images4bpp = BenchmarkHelper.GetSampleDataImages<ColorRGBA>();
     }
 
     #endregion
@@ -44,6 +49,26 @@ public class MinMaxBenchmarks
         IMinMax[] minMax = new IMinMax[_colors4bpp.Count];
         for (int i = 0; i < _colors4bpp.Count; i++)
             minMax[i] = new ReadOnlySpan<ColorRGBA>(_colors4bpp[i]).MinMax();
+
+        return minMax;
+    }
+
+    [Benchmark]
+    public IMinMax[] PixelHelper_3BPP_Image()
+    {
+        IMinMax[] minMax = new IMinMax[_images3bpp.Count];
+        for (int i = 0; i < _images3bpp.Count; i++)
+            minMax[i] = _images3bpp[i].MinMax();
+
+        return minMax;
+    }
+
+    [Benchmark]
+    public IMinMax[] PixelHelper_4BPP_Image()
+    {
+        IMinMax[] minMax = new IMinMax[_images4bpp.Count];
+        for (int i = 0; i < _images4bpp.Count; i++)
+            minMax[i] = _images4bpp[i].MinMax();
 
         return minMax;
     }