--- a/HPPH.Benchmark/SumBenchmarks.cs
+++ b/HPPH.Benchmark/SumBenchmarks.cs
@@ -13,6 +13,8 @@ public class SumBenchmarks
 
     private readonly List<ColorRGB[]> _colors3bpp;
     private readonly List<ColorRGBA[]> _colors4bpp;
+    private readonly List<IImage<ColorRGB>> _images3bpp;
+    private readonly List<IImage<ColorRGBA>> _images4bpp;
 
     #endregion
 
@@ -22,6 +24,9 @@ public class SumBenchmarks
     {
         _colors3bpp = BenchmarkHelper.GetSampleData<ColorRGB>();
         _colors4bpp = BenchmarkHelper.GetSampleData<ColorRGBA>();
+
+        _images3bpp = BenchmarkHelper.GetSampleDataImages<ColorRGB>();
+        _images4bpp = BenchmarkHelper.GetSampleDataImages<ColorRGBA>();
     }
 
     #endregion
@@ -44,6 +49,26 @@ public class SumBenchmarks
         ISum[] sums = new ISum[_colors4bpp.Count];
         for (int i = 0; i < _colors4bpp.Count; i++)
             sums[i] = new ReadOnlySpan<ColorRGBA>(_colors4bpp[i]).Sum();
+
+        return sums;
+    }
+
+    [Benchmark]
+    public ISum[] PixelHelper_3BPP_Image()
+    {
+        ISum[] sums = new ISum[_colors3bpp.Count];
+        for (int i = 0; i < _images3bpp.Count; i++)
+            sums[i] = _images3bpp[i].Sum();
+
+        return sums;
+    }
+
+    [Benchmark]
+    public ISum[] PixelHelper_4BPP_Image()
+    {
+        ISum[] sums = new ISum[_images4bpp.Count];
+        for (int i = 0; i < _images4bpp.Count; i++)
+            sums[i] = _images4bpp[i].Sum();
 
         return sums;
     }