--- a/HPPH.Generators/ColorFormatSourceGenerator.cs
+++ b/HPPH.Generators/ColorFormatSourceGenerator.cs
@@ -19,7 +19,6 @@ public class ColorSourceGenerator : IIncrementalGenerator
     private static readonly IGeneratorFeature[] FEATURES =
     [
         new Colors(),
-        new Average(),
         new MinMax(),
         new Sum(),
         new Quantize()