--- a/HPPH/Generated/HPPH.Generators/HPPH.Generators.ColorSourceGenerator/IColorFormat.MinMax.g.cs
+++ b/HPPH/Generated/HPPH.Generators/HPPH.Generators.ColorSourceGenerator/IColorFormat.MinMax.g.cs
@@ -4,5 +4,6 @@ namespace HPPH;
 
 public partial interface IColorFormat
 {
-    internal IMinMax MinMax(ReadOnlySpan<byte> data);
+    internal IMinMax ToMinMax(Generic3ByteMinMax data);
+    internal IMinMax ToMinMax(Generic4ByteMinMax data);
 }