--- a/HPPH/PixelHelper.cs
+++ b/HPPH/PixelHelper.cs
@@ -3,4 +3,7 @@
 namespace HPPH;
 
 [SkipLocalsInit]
-public static partial class PixelHelper;
+public static partial class PixelHelper
+{
+    private static readonly ParallelOptions PARALLEL_OPTIONS = new() { MaxDegreeOfParallelism = Environment.ProcessorCount };
+}