… HPPH.Generators/Features/Average.cs
… HPPH/PixelHelper.MinMax.cs
… HPPH/PixelHelper.Sum.cs