WIP

Results so far:
### Sum

| Method                     | Mean     | Error   | StdDev  | Allocated |
|--------------------------- |---------:|--------:|--------:|----------:|
| PixelHelper_3BPP           | 108.4 us | 1.04 us | 0.97 us |     528 B |
| PixelHelper_4BPP           | 172.7 us | 2.67 us | 2.49 us |     528 B |
| PixelHelper_3BPP_Image     | 157.4 us | 0.59 us | 0.52 us |     528 B |
| PixelHelper_4BPP_Image     | 210.0 us | 1.51 us | 1.34 us |     528 B |
| PixelHelper_3BPP_Old       | 108.1 us | 0.68 us | 0.57 us |     528 B |
| PixelHelper_4BPP_Old       | 172.8 us | 1.83 us | 1.62 us |     528 B |
| PixelHelper_3BPP_Image_Old | 293.7 us | 1.65 us | 1.47 us |     528 B |
| PixelHelper_4BPP_Image_Old | 472.9 us | 2.86 us | 2.23 us |     528 B |



### Average

It is boxing a few things which might be possible to avoid too, but it's still way better than before

| Method                     | Mean     | Error   | StdDev  | Allocated |
|--------------------------- |---------:|--------:|--------:|----------:|
| PixelHelper_3BPP           | 108.8 us | 0.72 us | 0.67 us |      56 B |
| PixelHelper_4BPP           | 170.2 us | 2.28 us | 2.13 us |      64 B |
| PixelHelper_3BPP_Image     | 154.8 us | 1.32 us | 1.24 us |     704 B |
| PixelHelper_4BPP_Image     | 213.4 us | 1.95 us | 1.83 us |     712 B |
| PixelHelper_3BPP_Old       | 109.0 us | 1.11 us | 1.04 us |      56 B |
| PixelHelper_4BPP_Old       | 173.0 us | 1.61 us | 1.51 us |      64 B |
| PixelHelper_3BPP_Image_Old | 296.0 us | 2.51 us | 2.23 us |      56 B |
| PixelHelper_4BPP_Image_Old | 507.1 us | 4.16 us | 4.08 us |      65 B |



### MinMax

| Method                     | Mean     | Error   | StdDev  | Allocated |
|--------------------------- |---------:|--------:|--------:|----------:|
| PixelHelper_3BPP           | 106.2 us | 0.16 us | 0.14 us |     312 B |
| PixelHelper_4BPP           | 138.6 us | 0.19 us | 0.17 us |     312 B |
| PixelHelper_3BPP_Image     | 203.3 us | 2.29 us | 2.15 us |     312 B |
| PixelHelper_4BPP_Image     | 241.8 us | 0.75 us | 0.66 us |     312 B |
| PixelHelper_3BPP_Old       | 130.9 us | 0.61 us | 0.51 us |     312 B |
| PixelHelper_4BPP_Old       | 140.8 us | 0.37 us | 0.31 us |     312 B |
| PixelHelper_3BPP_Image_Old | 289.1 us | 3.82 us | 3.58 us |     312 B |
| PixelHelper_4BPP_Image_Old | 411.0 us | 5.78 us | 5.40 us |     312 B |